Refactor code structure based on #80 PR comments
